pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "vdesk-manager"
version = "1.2.0"
description = "Windows 虚拟桌面管理工具"
requires-python = ">=3.8"
dependencies = [
    "pyvda",
    "pystray",
    "Pillow>=9.0",
    "pynput",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "windows: marks tests as Windows-specific",
]
